+++ hois_back/src/main/java/ee/hitsa/ois/web/dto/StateCurriculumModuleDto.java
@@ -18,7 +18,7 @@ public class StateCurriculumModuleDto extends VersionedCommand {
     private String module;  
     private String nameEt;
     private String nameEn;
-    private Integer credits;
+    private Double credits;
     private String objectivesEt;
     private String objectivesEn;
     private String assessmentsEt;
@@ -76,11 +76,11 @@ public class StateCurriculumModuleDto extends VersionedCommand {
         this.nameEn = nameEn;
     }
 
-    public Integer getCredits() {
+    public Double getCredits() {
         return credits;
     }
 
-    public void setCredits(Integer credits) {
+    public void setCredits(Double credits) {
         this.credits = credits;
     }
